data_storage/subscriber/requirements.txt
paho-mqtt
mysql-connector-python
requests

